[OS] Add update_frequency option when adding a task
Prepares for adding backlight management, where the OS manages dimming
of the display/glow LEDs based on a variable timeout. When the user
touches the display it will wake, optionally consuming the touch event
so no on-screen buttons are acidentally triggered.

Do do this effectively, without lag, we need the main run loop to
continue at full speed to process touch events. Specific tasks may want
to run at a more leisurely interval. If a specific task inserts a sleep,
then it would delay touch processing by the OS. We can't use async for
this, as the touch processing will need to consume the touch event and
re-poll, so that tasks don't see the wake touches. This seems like a
workable solution, that also saves code in userland if you do want
slower updates.

Im sure these tests will be a source of pain as they're introducing CPU
dependent results. Hopefully a 10% margin for error will be ok....