[OS] Add Task.touch_forces_execution
Adds support in the run loop to run tasks immediately if there is an
active touch, regardless of their preferred updated frequency.

This makes it much easier to add interactivity to slow update
interfaces, as you can leverage the OS scheduler to know when you need
to be called more freqently to handle the touch.