[BlkMgr] Fix issues with time_us wrapping waking the display
Time us wraps over long periods, but as we're doing maths on an interval
longer than the wrap period, there isn't any way to fix this.

The display is usually inactive for long periods of time, and the
timeouts are set in seconds, so switch the BacklightManager to using
seconds.

Considered using time.time_ns, but that is noted as having a heap
allocation. Probably not significant, but lets see how this goes, and
consider switching everything to time_ns once we have a baseline set of
functionality and can look in to the performance aspects of things.