[UI] Add WindoManager with basic page management
Bit of a big one, adds a WindowManager that allows registration of
multiple pages, one of which is currently active. Pages have their own
update frequency, managed by the main OS run loop.

This attempts to split out drawing into a "theme", that is independent
from any particular ui elements/controls. Hopefully this will let us
build a library of re-usable components, that can be re-skinned by
particular apps.

The separation of concerns here might not be quite right yet, but well
see as we build more on it.