[OS] Ensure touch_forces_execution causes run after touch ended
Previously, we didn't tick immediately after a touch ened. This resulted
in cleanup from a touch not occuring untill the next scheduled interval,
which could be a while off.

This generally reflects itself as stale UI state as redraw hasn't
happend. This adds a single extra execution immeidately after the touch
state returns to False, to give the task a chance to clean up.