[WM] Add `PushButton` and basic control management to Pages
This is probgably going to shuffle about a bit over time. It attemps to
simlify the management of on-screen controls that deliberately have side
effects, and decouple this logic from your main draw code.

I need to probably steal more from ImGUI, but at this point, a
semi-imperative approach still feels more aligned with the official
Presto SDK, so may be more familiar to people.

One of the main things here is an attempt to keep drawing within Theme,
without a theme being coupled to any specific controls, whilst also
allowing any custom user control class to override the default drawing
however it likes, without needing to modify the theme.

As such, the WM asks controls to draw themselves (to make it easier to
customise in end-user code), and the controls use well-defined theme
static methdods to draw text/frames etc.