[WM] Refactor font scaling mechanism
The old approach of allowing the absolute scale to be specified to a
theme's text calls made stable layouts difficult. If one theme had a
different default font scale to another, you'd end up with the relative
sizes of text changing.

This refactors the text handling so that a Theme defines a
base_font_scale, and all draw calls are then scaled relative to this.
This means a theme is then free to adjsut its sizing, and as long as
`Theme.line_spacing` is used by rendering code, on-screen text and its
positions should adjust proportionally.

This will be much more apparent when the upcoming vector text rendering
is used.